WIP: Now able to configure techsport

+++ backend/src/main/kotlin/com/techso/techsport/model/Team.kt
@@ -7,16 +7,16 @@ import kotlin.math.max
 open class Team(@Id val name: String) {
     var members: MutableList<String> = mutableListOf()
         private set;
-    var points: Long = 0
+    var points: Double = 0.0
         private set;
-    var pointsChanges: MutableMap<LocalDate, Long> = mutableMapOf()
+    var pointsChanges: MutableMap<LocalDate, Double> = mutableMapOf()
         private set;
 
-    fun addTime(points: Long) {
+    fun addTime(points: Double) {
         this.points += points
-        this.points = max(this.points, 0L);
+        this.points = max(this.points, 0.0);
 
-        val oldValue = this.pointsChanges.getOrDefault(LocalDate.now(), 0);
+        val oldValue = this.pointsChanges.getOrDefault(LocalDate.now(), 0.0);
         this.pointsChanges[LocalDate.now()] = oldValue + points
     }
 }
